split ValidationContext into a public file
Just minor tweaks to make it compile, moving the template functions to
the header and also marking virtuals as final.

In the next commits it will be moved to more public usage, including
the generator.
